requirements.txt
openenv-core>=0.2.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
openai>=1.30.0
pytest>=8.0.0